Add user, moderation, and report view modules
